daus comit 14

+++ app/src/main/java/com/maverick/kmjshowroom/Model/MobilDetailResponse.kt
@@ -1,3 +1,5 @@
+import java.math.BigInteger
+
 data class MobilDetailResponse(
     val code: Int,
     val mobil: MobilData,
@@ -9,12 +11,12 @@ data class MobilData(
     val kode_mobil: String,
     val kode_user: String,
     val nama_mobil: String,
-    val tahun_mobil: Int,        // ✅ ubah dari String ke Int
-    val jarak_tempuh: Int,       // ✅ ubah dari String ke Int
-    val full_prize: Int,         // ✅ ubah dari String ke Int
-    val uang_muka: Int,          // ✅ ubah dari String ke Int
-    val tenor: Int,              // ✅ ubah dari String ke Int
-    val angsuran: Int,           // ✅ ubah dari String ke Int
+    val tahun_mobil: Int,
+    val jarak_tempuh: Int,
+    val full_prize: BigInteger,
+    val uang_muka: Int,
+    val tenor: Int,
+    val angsuran: Int,
     val jenis_kendaraan: String,
     val sistem_penggerak: String,
     val tipe_bahan_bakar: String,
