+++ app/src/main/java/com/maverick/kmjshowroom/Model/MobilListResponse.kt
@@ -1,5 +1,7 @@
 package com.maverick.kmjshowroom.Model
 
+import java.math.BigInteger
+
 data class MobilListResponse(
     val code: Int,
     val data: List<MobilItem>
@@ -12,7 +14,7 @@ data class MobilItem(
     val warna_exterior: String,
     val tipe_bahan_bakar: String,
     val jarak_tempuh: Int,
-    val full_prize: Int,
+    val full_prize: BigInteger,
     val angsuran: Int,
     val tenor: Int,
     val dp: Int,
